+++ tests/models.py
@@ -1,5 +1,6 @@
+from django.contrib.contenttypes.fields import GenericForeignKey
+from django.contrib.contenttypes.models import ContentType
 from django.db import models
-from django.urls import reverse_lazy
 
 from django.utils.translation import ugettext_lazy as _
 
@@ -37,6 +38,11 @@ class SimpleSales(models.Model):
     value = models.DecimalField(_('value'), max_digits=19, decimal_places=2, default=0)
     created_at = models.DateTimeField(null=True, verbose_name=_('Created at'))
     flag = models.CharField(max_length=50, default='sales')
+
+    content_type = models.ForeignKey(ContentType, on_delete=models.DO_NOTHING, null=True)
+    object_id = models.PositiveIntegerField(null=True)
+    content_object = GenericForeignKey('content_type', 'object_id')
+
 
     def save(self, force_insert=False, force_update=False, using=None, update_fields=None):
         self.value = self.quantity * self.price