+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [0.5.1] - 
 
 - Added tests for group by a flag field
-
+- Fixed an issue if GenericForeignKey is on the report model
 
 
 ## [0.5.0] - 2020-12-11